Implement Eq and Hash for GreenTokenData
`GreenToken` implements all traits required to use it as the key
type of a hash map. It also implements `Borrow<GreenTokenData>`.
Thus is would be nice if `GreenTokenData` also implements `Eq` and
`Hash` such that the result of `SyntaxToken::green` can be used for
lookups in such a hash map without additional overhead in terms of
sytntax and runtime.